requirements.txt
# Core scientific computing
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Deep learning frameworks
torch>=1.9.0
torchvision>=0.10.0

# Quantum computing (optional)
pennylane>=0.23.0

# Data processing and visualization
matplotlib>=3.4.0
seaborn>=0.11.0
tqdm>=4.62.0
opencv-python>=4.5.0

# Configuration and serialization
PyYAML>=6.0
jsonschema>=4.0.0

# Utilities
requests>=2.25.0
h5py>=3.4.0
pillow>=8.3.0

# Development tools
pytest>=6.2.0
black>=21.0.0
flake8>=3.9.0
mypy>=0.910

# Jupyter support (optional)
jupyter>=1.0.0
notebook>=6.4.0
